Updated the linker to use size_t type data_metric. I also modified linker.pyx to use the long int versions of the ObjGen library functions. This *should* be enough to enable massive datacube processing.

+++ src/linker/RJJ_ObjGen_Dmetric.cpp
@@ -1,9 +1,9 @@
 #include<iostream>
-#include "RJJ_ObjGen.h"
+#include<RJJ_ObjGen.h>
 
 using namespace std;
 
-void CreateMetric(int * data_metric, int * xyz_order, int NOx, int NOy, int NOz){
+void CreateMetric(size_t * data_metric, int * xyz_order, int NOx, int NOy, int NOz){
 
   // ensure that the x_order, y_order and z_order values are 1,2,3
   if((xyz_order[0] <= xyz_order[1]) && (xyz_order[0] <= xyz_order[2])){
